requirements.txt
fastapi
jinja2
openai
opencc-python-reimplemented
pydantic
python-docx
python-multipart
uvicorn
